put statusbar progress bar to gui object
The problem here is that statusBar().removeWidget() not really removes
the widget from the statusbar, but only hides it (see documentation).
Therefore we need it makes more sense to keep a common statusbar
progressbar, and pass it to the background loader.  To activate it
one has to call both addWidget and show according to the documentation.